+++ tests/fibonacci_numbers/test_efficient_solution.py
@@ -5,7 +5,7 @@ import pytest
 
 from solutions import fibonacci_numbers
 from tests.fibonacci_numbers import StressTestsParams
-from tests.measure_performance import PerformanceMeasures
+from tests.performance_measures import PerformanceMeasures
 from tests.read_config import get_names, read_config
 
 """ Performance tests show that the efficient solution is 10000 times faster for
