+++ tests/greatest_common_divisor/test_efficient_solution.py
@@ -5,7 +5,7 @@ import pytest
 
 from solutions import greatest_common_divisor
 from tests.greatest_common_divisor.stress_tests_params import StressTestsParams
-from tests.measure_performance import PerformanceMeasures
+from tests.performance_measures import PerformanceMeasures
 from tests.read_config import get_names, read_config
 
 """ Performance tests show that the efficient solution is a 100 times faster for
